feat: using bind mount to read config file
To apply changes to the config file, the image had to be re-built.
Now, a bind mount is created at /config in the container, so that
it just has to be restarted.
Later on, we might want to be able to send a 'reload' signal to
the container, thus eliminating the need to restart the container.